requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.0=py312h06a4308_0
aiohttp=3.10.5=py312h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
annotated-types=0.6.0=py312h06a4308_0
anyio=4.6.2=py312h06a4308_0
attrs=24.2.0=py312h06a4308_0
blinker=1.6.2=py312h06a4308_0
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
cachetools=5.3.3=py312h06a4308_0
certifi=2024.8.30=py312h06a4308_0
cffi=1.17.1=py312h1fdaa30_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py312h06a4308_0
cryptography=43.0.0=py312hdda0065_0
distro=1.9.0=py312h06a4308_0
expat=2.6.3=h6a678d5_0
frozenlist=1.4.0=py312h5eee18b_0
google-api-core=2.19.1=py312h06a4308_0
google-api-python-client=2.149.0=pyhff2d567_0
google-auth=2.29.0=py312h06a4308_0
google-auth-httplib2=0.2.0=pyhd8ed1ab_0
google-auth-oauthlib=0.5.2=py312h06a4308_0
googleapis-common-protos=1.63.2=py312h06a4308_0
h11=0.14.0=py312h06a4308_0
httpcore=1.0.2=py312h06a4308_0
httplib2=0.22.0=pyhd8ed1ab_0
httpx=0.27.0=py312h06a4308_0
idna=3.7=py312h06a4308_0
iniconfig=1.1.1=pyhd3eb1b0_0
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20240116.2=cxx17_h6a678d5_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libprotobuf=4.25.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
multidict=6.0.4=py312h5eee18b_0
ncurses=6.4=h6a678d5_0
notion-client=2.2.1=pyhd8ed1ab_0
oauthlib=3.2.2=py312h06a4308_0
openai=1.37.0=py312h06a4308_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=py312h06a4308_0
pip=24.2=py312h06a4308_0
pluggy=1.0.0=py312h06a4308_1
proto-plus=1.24.0=py312h06a4308_0
protobuf=4.25.3=py312h12ddb61_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.8.2=py312h06a4308_0
pydantic-core=2.20.1=py312hb02cf49_0
pyjwt=2.8.0=py312h06a4308_0
pyopenssl=24.2.1=py312h06a4308_0
pyparsing=3.1.2=py312h06a4308_0
pysocks=1.7.1=py312h06a4308_0
pytest=7.4.4=py312h06a4308_0
python=3.12.7=h5148396_0
python-dotenv=0.21.0=py312h06a4308_0
readline=8.2=h5eee18b_0
requests=2.32.3=py312h06a4308_0
requests-oauthlib=2.0.0=py312h06a4308_0
rsa=4.7.2=pyhd3eb1b0_1
setuptools=75.1.0=py312h06a4308_0
sniffio=1.3.0=py312h06a4308_0
sqlite=3.45.3=h5eee18b_0
tk=8.6.14=h39e8969_0
tqdm=4.66.5=py312he106c6f_0
typing-extensions=4.11.0=py312h06a4308_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024b=h04d1e81_0
uritemplate=4.1.1=pyhd8ed1ab_0
urllib3=2.2.3=py312h06a4308_0
wheel=0.44.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
yarl=1.11.0=py312h5eee18b_0
zlib=1.2.13=h5eee18b_1
